backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
openai
pandas
scikit-learn